backend/pyproject.toml
[tool.poetry]
name = "charts"
version = "0.1.0"
description = ""
authors = ["Edgar Pabon <edgar.pabon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
django-cors-headers = "^4.9.0"
djangorestframework = "^3.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
